Separate disk handling code into DiskHandlerMixin
