+++ astoria/managers/mixins/__init__.py
@@ -1,0 +1,1 @@
+"""Manager Mixins."""
